# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MiniStay API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/rooms/{room}": {
            "get": {
                "tags": [
                    "User - Rooms"
                ],
                "summary": "Get room detail",
                "operationId": "701b922d8c0459c9b3e62dd31b62edc9",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "date": {
                                            "$ref": "#/components/schemas/Room"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rooms/{room}/calendar": {
            "get": {
                "tags": [
                    "User - Rooms"
                ],
                "summary": "Get availability calendar for a room",
                "operationId": "04d0500806b4f5fd0431bedbf6a07b82",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room availability calendar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "room": {
                                                    "type": "integer"
                                                },
                                                "calendar": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/RoomAvailability"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rooms/{room}/book": {
            "post": {
                "tags": [
                    "User - Booking"
                ],
                "summary": "Create booking for a room",
                "operationId": "0aa095c8545c703d647a9a0be0c2dbff",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "check_in",
                                    "check_out"
                                ],
                                "properties": {
                                    "check_in": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "check_out": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Booking"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Room unavailable or validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/my-bookings": {
            "get": {
                "tags": [
                    "User - Booking"
                ],
                "summary": "Get user's bookings",
                "operationId": "f1639ec7fe2b4b6cc818bc001c0c128a",
                "responses": {
                    "200": {
                        "description": "List of user's bookings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Booking"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/bookings/{booking}/review": {
            "post": {
                "tags": [
                    "User - Reviews"
                ],
                "summary": "Submit review for a completed booking",
                "operationId": "a8c1dafde870eeb5ab8bf04069a06899",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "description": "Booking ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "comment": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Review created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Review"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/my-reviews": {
            "get": {
                "tags": [
                    "User - Reviews"
                ],
                "summary": "Get reviews created by authenticated user",
                "operationId": "455807dd9dd12b4e8137e5082784d00b",
                "responses": {
                    "200": {
                        "description": "List of user's reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Review"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Booking": {
                "required": [
                    "user_id",
                    "room_id",
                    "check_in",
                    "check_out",
                    "total_price",
                    "booking_code"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "room_id": {
                        "type": "integer"
                    },
                    "check_in": {
                        "type": "string",
                        "format": "date"
                    },
                    "check_out": {
                        "type": "string",
                        "format": "date"
                    },
                    "total_price": {
                        "type": "integer"
                    },
                    "booking_code": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "waiting_payment",
                            "paid",
                            "cancelled",
                            "checked_in",
                            "completed"
                        ]
                    },
                    "qr_code_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Review": {
                "required": [
                    "booking_id",
                    "user_id",
                    "room_id",
                    "rating"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "booking_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "room_id": {
                        "type": "integer"
                    },
                    "rating": {
                        "type": "integer"
                    },
                    "comment": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_visible": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Room": {
                "required": [
                    "name",
                    "price_per_day"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "price_per_day": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "facilities": {
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "RoomAvailability": {
                "required": [
                    "room_id",
                    "date",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "room_id": {
                        "type": "integer"
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "available",
                            "booked"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}